requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
psycopg2
ollama
matplotlib